hmily

+++ Week_08/hmily-transaction/hmily-order/src/main/java/org/fzb/order/controller/OrderController.java
@@ -6,6 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigInteger;
+
 /**
  * OrderController
  *
@@ -19,7 +21,7 @@ public class OrderController {
     private OrderService orderService;
 
     @GetMapping("/create")
-    public int createOrder(){
+    public BigInteger createOrder(){
         return  orderService.createOrder();
     }
 }
